fix(200): eight fences moved with stories 02/03 and #526, paid without a budget raised — the build ledger regenerated from its generator (Phase 200 in the phase index); web_runtime.py back under the boot-only density budget by carving the identity capture, the database claim and the sweep gate into holdspeak/runtime/ownership.py (DatabaseOwnershipMixin; the guard's own prescribed seam; documented); a dash in RELEASING's prose; the ask v1 contract hash re-derived as a DELIBERATE BUMP with the reason (PR #526 made the grounding echo always a dict; red on main since Sept 1; the field set unchanged, the schema version not bumped, for review); the two phase-143 censuses re-anchored from their authorities (no new inference site, resolver or pointer); one-path cardinality's cancellation test diagnosed as a pre-existing scheduling race (ledgered in 158; nothing widened); the cadence-routes setup error diagnosed as a fixture racing the database singleton (a fixture fix, ledgered)
PMO-Story: HS-200-05
PMO-Contract-Digest: sha256:26287f5245f548f8ae8447d696f60574915813aa8d671e1876884e06a155ad34
