+++ tests/unit/test_phase143_routing_authority_census.py
@@ -55,17 +55,17 @@ ROUTING_RESOLVER_NAMES = frozenset({
 ROUTING_RESOLVER_DEFINITIONS = {
     "holdspeak/deployment_revisions.py:206:resolve_workbench_deployment_revision",
     "holdspeak/deployment_revisions.py:228:resolve_deployment_revision",
-    "holdspeak/inference_targets.py:550:resolve_placement",
-    "holdspeak/inference_targets.py:496:resolve_inference_target",
-    "holdspeak/inference_targets.py:590:resolve_thought_placement",
+    "holdspeak/inference_targets.py:567:resolve_placement",
+    "holdspeak/inference_targets.py:513:resolve_inference_target",
+    "holdspeak/inference_targets.py:607:resolve_thought_placement",
     "holdspeak/intel/providers.py:666:resolve_meeting_placement",
 }
 
 ROUTING_RESOLVER_REFERENCES = {
     "holdspeak/deployment_revisions.py:209:import:resolve_inference_target",
     "holdspeak/deployment_revisions.py:218:ref:resolve_inference_target",
-    "holdspeak/inference_targets.py:584:ref:resolve_inference_target",
-    "holdspeak/inference_targets.py:601:ref:resolve_placement",
+    "holdspeak/inference_targets.py:601:ref:resolve_inference_target",
+    "holdspeak/inference_targets.py:618:ref:resolve_placement",
     "holdspeak/intel/__init__.py:63:import:resolve_meeting_placement",
     "holdspeak/intel/providers.py:193:ref:resolve_meeting_placement",
     "holdspeak/intel/providers.py:235:ref:resolve_meeting_placement",
@@ -75,8 +75,8 @@ ROUTING_RESOLVER_REFERENCES = {
     "holdspeak/kernel/inference_invoke.py:92:ref:resolve_deployment_revision",
     "holdspeak/kernel/inference_runner.py:13:import:resolve_deployment_revision",
     "holdspeak/kernel/inference_runner.py:851:ref:resolve_deployment_revision",
-    "holdspeak/services/ask_service.py:307:import:resolve_placement",
-    "holdspeak/services/ask_service.py:308:ref:resolve_placement",
+    "holdspeak/services/ask_service.py:309:import:resolve_placement",
+    "holdspeak/services/ask_service.py:310:ref:resolve_placement",
     "holdspeak/services/inference_setup_service.py:184:ref:resolve_inference_target",
     "holdspeak/services/inference_setup_service.py:23:import:resolve_inference_target",
     "holdspeak/services/model_profile_service.py:688:import:resolve_inference_target",
@@ -160,7 +160,7 @@ PROFILE_ID_CLASSIFICATIONS = {
         "holdspeak/web/routes/front_door.py:416:profile_id",
         "holdspeak/web/routes/front_door.py:417:profile_id",
         "holdspeak/web/routes/front_door.py:444:profile_id",
-        "holdspeak/services/ask_service.py:318:profile_id",
+        "holdspeak/services/ask_service.py:320:profile_id",
         "holdspeak/services/inference_setup_service.py:653:profile_id", "holdspeak/services/settings_service.py:103:profile_id",
         "holdspeak/setup_status.py:151:profile_id",
         "holdspeak/services/model_profile_service.py:225:profile_id",
@@ -176,7 +176,7 @@ PROFILE_ID_CLASSIFICATIONS = {
     **{site: "migration source" for site in {
         "holdspeak/db/models/__init__.py:1136:profile_id",
         "holdspeak/db/models/workbench.py:138:profile_id",
-        "holdspeak/services/recipe_service.py:364:profile_id",
+        "holdspeak/services/recipe_service.py:416:profile_id",
         "holdspeak/services/workbench_service.py:561:profile_id",
     }},
     **{site: "credential/provider identity" for site in {
