+++ uat/features.yaml
@@ -15,7 +15,7 @@ source:
   inventory: pm/roadmap/holdspeak-uat/phase-2-the-inventory/directory/_raw-inventory-rows.json
   additions: uat/feature-additions.yaml
   phase_index: pm/roadmap/holdspeak/README.md
-  phases_total: 131
+  phases_total: 132
 features:
 - key: activity.candidates.meeting
   title: Activity-derived meeting candidates
@@ -4346,4 +4346,6 @@ phase_map:
   '179':
   - internal/no-uat-surface
   '180':
+  - internal/no-uat-surface
+  '200':
   - internal/no-uat-surface
